move wrappers to a specific module
